Index: Translation
I plan to get back to this work, but I do not want to continue it in
Italian. I prefer to switch it to English.

This commit translates the main parts and chapters titles in English.
Last part titles have not been translated for now. It is likely that I
will remove or move the material that is in there, so it does not make
any sense to translate anything that is in there for now.